Removed coordinate system, interaction and configuration now happens through cities only Added docker configuration for easy building and testing (and future GHCR integration) Added a .env.sample file Tweaked configuration files to reflect latest changes Moved getConfigValue() over to PluginConfig.java
