+++ papermc-mod/.env.sample
@@ -1,0 +1,4 @@
+SKYLINK_HOST=localhost
+SKYLINK_PORT=8080
+
+SKYLINK_CITY=UK/London
