+++ papermc-mod/src/main/java/skylink/domain/config/PluginConfig.java
@@ -7,28 +7,26 @@ public class PluginConfig {
     private final int port;
 
     private final String defaultCity;
-
-    private final double defaultLatitude;
-    private final double defaultLongitude;
-
     private SkyLink plugin;
 
-    public PluginConfig(SkyLink plugin) {
-        this.plugin = plugin;
+    public PluginConfig() {
+        this.host = getConfigValue("SKYLINK_HOST", "config.host", "localhost");
+        this.port = Integer.parseInt(getConfigValue("SKYLINK_PORT", "config.port", "8080"));
+        this.defaultCity = getConfigValue("SKYLINK_CITY", "location.city", "UK/London");
+    }
 
-        this.host = plugin.getConfigValue("SKYLINK_HOST", "config.host", "localhost");
-        this.port = Integer.parseInt(plugin.getConfigValue("SKYLINK_PORT", "config.port", "8080"));
+    public String getConfigValue(String envKey, String ymlKey, String defaultValue) {
+        String envValue = System.getenv(envKey);
+        if (envValue != null && !envValue.isEmpty()) return envValue;
 
-        this.defaultCity = plugin.getConfigValue("SKYLINK_CITY", "location.city", "UK/London");
-        this.defaultLatitude = Double.parseDouble(plugin.getConfigValue("SKYLINK_LATITUDE", "location.latitude", "0"));
-        this.defaultLongitude = Double.parseDouble(plugin.getConfigValue("SKYLINK_LONGITUDE", "location.longitude", "0"));
+        String ymlValue = plugin.getConfig().getString(ymlKey);
+        if (ymlValue != null && !ymlValue.isEmpty()) return ymlValue;
+
+        return defaultValue;
     }
 
     public String getHost() { return host; }
     public int getPort() { return port; }
 
     public String getDefaultCity() { return defaultCity; }
-
-    public double getDefaultLatitude() { return defaultLatitude; }
-    public double getDefaultLongitude() { return defaultLongitude; }
 }
