+++ papermc-mod/src/main/resources/config.yml
@@ -4,5 +4,3 @@ config:
 
 location:
   city: "UK/London"
-  latitude: 0
-  longitude: 0
