+++ papermc-mod/src/main/resources/plugin.yml
@@ -8,7 +8,7 @@ description: Have real world weather in Minecraft!
 commands:
   skylink:
     description: Use SkyLink to manipulate Minecraft weather with real world data
-    usage: /skylink <city|coordinates> <value>
+    usage: /skylink city <Country/City>
     permission: skylink.use
 
 permissions:
